backend/pyproject.toml
[project]
name = "teamfit-backend"
version = "0.1.0"
description = "TEAMFIT MVP - AI-powered team-building platform backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "supabase",
    "svix",
    "pydantic",
    "python-multipart",
    "PyJWT",
    "openai>=1.3.5",
    "pypdf>=3.17.1",
    "pdfplumber>=0.10.3",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.2",
    "httpx>=0.25.2",
    "pydantic-settings>=2.0.0",
    "celery>=5.3.4",
    "redis>=5.0.1",
]
